feat(toolbox): moved setMainToolbarThresholds to general actions
